Struct lyon_bezier::QuadraticBezierSegment

pub struct QuadraticBezierSegment {
    pub from: Vec2,
    pub ctrl: Vec2,
    pub to: Vec2,
}

A 2d curve segment defined by three points: the beginning of the segment, a control point and the end of the segment.

The curve is defined by equation: ∀ t ∈ [0..1], P(t) = (1 - t)² * from + 2 * (1 - t) * t * ctrl + 2 * t * to

Fields
from: Vec2 ctrl: Vec2 to: Vec2
Methods

impl QuadraticBezierSegment

fn sample(&self, t: f32) -> Point

Sample the curve at t (expecting t between 0 and 1).

fn sample_x(&self, t: f32) -> f32

Sample the x coordinate of the curve at t (expecting t between 0 and 1).

fn sample_y(&self, t: f32) -> f32

Sample the y coordinate of the curve at t (expecting t between 0 and 1).

fn flip(&mut self)

Swap the beginning and the end of the segment.

fn find_y_maximum(&self) -> f32

Find the advancement of the y-most position in the curve.

This returns the advancement along the curve, not the actual y position.

fn find_y_inflection(&self) -> Option<f32>

Return the y inflection point or None if this curve is y-monotone.

fn split(&self, t: f32) -> (QuadraticBezierSegment, QuadraticBezierSegment)

Split this curve into two sub-curves.

fn before_split(&self, t: f32) -> QuadraticBezierSegment

Return the curve before the split point.

fn after_split(&self, t: f32) -> QuadraticBezierSegment

Return the curve after the split point.

fn to_cubic(&self) -> CubicBezierSegment

Elevate this curve to a third order bezier.

fn flattening_step(&self, tolerance: f32) -> f32

Find the interval of the begining of the curve that can be approximated with a line segment.

fn flattened_for_each<F: FnMut(Point)>(&self, tolerance: f32, call_back: &mut F)

Iterates through the curve invoking a callback at each point.

fn flattening_iter(&self, tolerance: f32) -> QuadraticFlatteningIter

Returns the flattened representation of the curve as an iterator, starting after the current point.